requirements.txt
# Social Media Posting Requirements
tweepy>=4.14.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Existing requirements (if any)
openai>=1.3.0